moved RAG related code into the utils file
